move TLSSocketFactory and trySend extension
