+++ rtmp/src/main/java/com/pedro/rtmp/utils/Utils.kt
@@ -16,25 +16,14 @@
 
 package com.pedro.rtmp.utils
 
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
 import java.io.InputStream
 import java.io.OutputStream
 import java.nio.ByteBuffer
-import java.util.concurrent.BlockingQueue
 
 /**
  * Created by pedro on 20/04/21.
  */
 
-inline infix fun <reified T: Any> BlockingQueue<T>.trySend(item: T): Boolean {
-  return try {
-    this.add(item)
-    true
-  } catch (e: IllegalStateException) {
-    false
-  }
-}
 
 fun InputStream.readUntil(byteArray: ByteArray) {
   var bytesRead = 0
