+++ rtsp/src/main/java/com/pedro/rtsp/utils/Extensions.kt
@@ -17,19 +17,8 @@
 package com.pedro.rtsp.utils
 
 import android.util.Base64
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
 import java.nio.ByteBuffer
-import java.util.concurrent.BlockingQueue
 
-inline infix fun <reified T: Any> BlockingQueue<T>.trySend(item: T): Boolean {
-  return try {
-    this.add(item)
-    true
-  } catch (e: IllegalStateException) {
-    false
-  }
-}
 
 fun ByteArray.encodeToString(flags: Int = Base64.NO_WRAP): String {
   return Base64.encodeToString(this, flags)
